# apps/api/src/promotions/dto/update-promotion.dto.ts
import {
  IsBoolean,
  IsDateString,
  IsEnum,
  IsNumber,
  IsOptional,
  IsString,
  MaxLength,
  Min,
  MinLength,
  Validate,
} from 'class-validator';
import { PromotionType } from '@prisma/client';
import { Type } from 'class-transformer';
import { PromotionDateWindowConstraint } from './promotion-date-window.constraint.js';

export class UpdatePromotionDto {
  @IsOptional()
  @IsString()
  @MinLength(2)
  @MaxLength(120)
  name?: string;

  @IsOptional()
  @IsEnum(PromotionType)
  type?: PromotionType;

  @IsOptional()
  @IsNumber()
  @Min(0)
  @Type(() => Number)
  value?: number | null;

  @IsOptional()
  @IsDateString()
  startsAt?: string | null;

  @IsOptional()
  @IsDateString()
  @Validate(PromotionDateWindowConstraint)
  endsAt?: string | null;

  @IsOptional()
  @IsBoolean()
  isActive?: boolean;
}